requirements.txt
# Basic functionality
numpy
pandas
# Machine Learning
tensorflow
opencv-python
# API backend
fastapi
python-multipart
requests
uvicorn
#Cloud
google-cloud-storage
